+++ client/src/input/BlockTypeTab.js
@@ -2,8 +2,6 @@ import Garnish from 'garnish'
 
 const _defaults = {
   name: '',
-  bodyHtml: '',
-  footHtml: '',
   errors: []
 }
 
@@ -13,32 +11,10 @@ export default Garnish.Base.extend({
     settings = Object.assign({}, _defaults, settings)
 
     this._name = settings.name
-    this._bodyHtml = settings.bodyHtml || ''
-    this._footHtml = settings.footHtml || ''
     this._errors = settings.errors
   },
 
   getErrors () { return Array.from(this._errors) },
 
-  getName () { return this._name },
-
-  getBodyHtml (blockId = null) {
-    return this._getHtml(this._bodyHtml, blockId)
-  },
-
-  getFootHtml (blockId = null) {
-    return this._getHtml(this._footHtml, blockId)
-  },
-
-  isBlank () {
-    return !this._bodyHtml.trim()
-  },
-
-  _getHtml (html, blockId = null) {
-    if (blockId !== null) {
-      return html.replace(/__NEOBLOCK__/g, blockId)
-    }
-
-    return html
-  }
+  getName () { return this._name }
 })